requirements.txt
apistar==0.3.9
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
py==1.4.34
pytest==3.2.3
PyYAML==3.12
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.12.2
whitenoise==3.3.1
